+++ CHANGES.md
@@ -5,6 +5,7 @@
 - [pull #483] Fix hashing nested HTML blocks
 - [pull #486] Fix backslash being unable to escape raw HTML tags
 - [pull #482] Add support for telegram spoiler in extras
+- [pull #485] mermaid support
 
 
 ## python-markdown2 2.4.6