+++ CHANGES.md
@@ -6,6 +6,8 @@
 - [pull #486] Fix backslash being unable to escape raw HTML tags
 - [pull #482] Add support for telegram spoiler in extras
 - [pull #485] mermaid support
+- [pull #490] Fix indented codeblocks inside fences (#489)
+- [pull #490] Remove `code-color` extra
 
 
 ## python-markdown2 2.4.6